Make use of app config vars
