+++ team2/server/IdeaJarAPI/WebAPI/Controllers/AuthController.cs
@@ -71,7 +71,7 @@ namespace WebAPI.Controllers
             if (!result.IsSuccess)
                 return BadRequest(result);
 
-            return Redirect($"{_configuration["AppUrl"]}/ConfirmEmail.html");
+            return Redirect($"{_configuration["APP_URL"]}/ConfirmEmail.html");
         }
 
         [HttpPost("ForgotPassword")]
